Migrate fjs/fsm to .f.mjs
Dependency-leaf implementation and proof migrate from authored
TypeScript to JSDoc-typed JavaScript under the stage-1 migration.
Grammar stays as a JSDoc typedef in the module — no external importers
of the type, so no sibling types.ts is needed. Updates two stale
module.f.ts doc references (fjs/types/function/operator/types.ts,
fjs/fsm/todo/scan-as-map.md).
